requirements.txt
# Core dependencies
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.1
groq==0.11.0

# Data validation
pydantic==2.9.2

# UI
streamlit==1.39.0

# Utilities
python-dotenv==1.0.1
typing-extensions==4.12.2

# Observability
langfuse==2.53.7
